+++ lib/pghero/methods/system.rb
@@ -5,9 +5,8 @@ module PgHero
         !system_stats_provider.nil?
       end
 
-      # TODO remove defined checks in 3.0
       def system_stats_provider
-        if aws_db_instance_identifier && (defined?(Aws) || defined?(AWS))
+        if aws_db_instance_identifier
           :aws
         elsif gcp_database_id
           :gcp
@@ -42,18 +41,13 @@ module PgHero
 
       def rds_stats(metric_name, duration: nil, period: nil, offset: nil, series: false)
         if system_stats_enabled?
-          aws_options = {region: region}
-          if access_key_id
-            aws_options[:access_key_id] = access_key_id
-            aws_options[:secret_access_key] = secret_access_key
+          aws_options = {region: aws_region}
+          if aws_access_key_id
+            aws_options[:access_key_id] = aws_access_key_id
+            aws_options[:secret_access_key] = aws_secret_access_key
           end
 
-          client =
-            if defined?(Aws)
-              Aws::CloudWatch::Client.new(aws_options)
-            else
-              AWS::CloudWatch.new(aws_options).client
-            end
+          client = Aws::CloudWatch::Client.new(aws_options)
 
           duration = (duration || 1.hour).to_i
           period = (period || 1.minute).to_i
@@ -108,7 +102,8 @@ module PgHero
           end
 
         client = Azure::Monitor::Profiles::Latest::Mgmt::Client.new
-        timespan = "#{start_time.iso8601}/#{end_time.iso8601}"
+        # call utc to convert +00:00 to Z
+        timespan = "#{start_time.utc.iso8601}/#{end_time.utc.iso8601}"
         results = client.metrics.list(
           azure_resource_id,
           metricnames: metric_name,
@@ -275,12 +270,13 @@ module PgHero
             used = azure_stats("storage_used", **options)
             free_space(quota, used)
           else
-            # no read_iops, write_iops
-            # could add io_consumption_percent
+            replication_lag_stat = azure_flexible_server? ? "physical_replication_delay_in_seconds" : "pg_replica_log_delay_in_seconds"
             metrics = {
               cpu: "cpu_percent",
               connections: "active_connections",
-              replication_lag: "pg_replica_log_delay_in_seconds"
+              replication_lag: replication_lag_stat,
+              read_iops: "read_iops", # flexible server only
+              write_iops: "write_iops" # flexible server only
             }
             raise Error, "Metric not supported" unless metrics[metric_key]
             azure_stats(metrics[metric_key], **options)
@@ -288,6 +284,10 @@ module PgHero
         else
           raise NotEnabled, "System stats not enabled"
         end
+      end
+
+      def azure_flexible_server?
+        azure_resource_id.include?("/Microsoft.DBforPostgreSQL/flexibleServers/")
       end
 
       # only use data points included in both series